requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
python-docx==0.8.11
openpyxl==3.0.9
SQLAlchemy==1.4.23
gunicorn==20.1.0
python-dotenv==0.19.0
